requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==2.7.1
django-js-asset==2.0.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.0
psycopg2-binary==2.9.5
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tablib==3.1.0
typing-extensions==4.1.1
xlrd==2.0.1
xlwt==1.3.0

